guided-init clean up
